+++ apps/feedback/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class FeedbackConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "apps.feedback"
+    verbose_name = "Обратная информация"
