+++ apps/feedback/models/__init__.py
@@ -1,0 +1,3 @@
+from .question import Question
+
+__all__ = (Question,)
