+++ apps/feedback/utilities.py
@@ -1,0 +1,12 @@
+from django.utils import timezone
+
+
+def get_festival_year():
+    if 7 <= timezone.now().month <= 12:
+        return timezone.now().year + 1
+    return timezone.now().year
+
+
+def generate_upload_path(instance, filename):
+    festival_year = get_festival_year()
+    return f"{instance.__class__.__name__}/{festival_year}/{filename}"
