+++ apps/feedback/views/__init__.py
@@ -1,0 +1,3 @@
+from .question import QuestionCreateAPIView
+
+__all__ = (QuestionCreateAPIView,)
