+++ apps/info/models/__init__.py
@@ -1,7 +1,6 @@
 from .festival import Festival, FestivalTeamMember, PressRelease
 from .people import Partner, Selector, Sponsor, Volunteer
 from .place import Place
-from .question import Question
 
 __all__ = (
     FestivalTeamMember,
@@ -10,7 +9,6 @@ __all__ = (
     Place,
     Sponsor,
     Volunteer,
-    Question,
     PressRelease,
     Selector,
 )
