+++ apps/info/utils.py
@@ -1,35 +1,8 @@
-from django.conf import settings
-from django.core.mail import EmailMessage
 from django.http import HttpResponse
 from django.template.loader import get_template
 from rest_framework import status
 from rest_framework.response import Response
 from xhtml2pdf import pisa
-
-from apps.core.models import Setting
-from apps.core.utils import send_email
-
-
-def send_question(instance):
-    """Transmit question parameters in the template.
-
-    For more information please read the documentation:
-    https://anymail.readthedocs.io/en/stable/esps/mailjet//
-    """
-    message = EmailMessage(
-        subject=Setting.get_setting("email_subject_for_question"),
-        from_email=Setting.get_setting("email_send_from"),
-        to=(Setting.get_setting("email_send_to"),),
-    )
-
-    message.template_id = settings.MAILJET_TEMPLATE_ID_QUESTION
-    message.merge_global_data = {
-        "question": instance.question,
-        "author_name": instance.author_name,
-        "author_email": instance.author_email,
-    }
-
-    return send_email(message)
 
 
 def get_pdf_response(press_release_instance, path_to_font):
