+++ config/settings/base.py
@@ -56,6 +56,7 @@ LOCAL_APPS = [
     "apps.articles",
     "apps.info",
     "apps.content_pages",
+    "apps.feedback",
 ]
 # https://docs.djangoproject.com/en/dev/ref/settings/#installed-apps
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS
